improve table.md docs
* add example with more markdown content
* rename table_markdown to avoid misunderstanding
* update tour.conf
