+++ py/examples/tour.conf
@@ -76,7 +76,6 @@ form_template.py
 form_markup.py
 stepper.py
 table_markdown.py
-table_markdown_links.py
 table_markdown_pandas.py
 table.py
 table_sort.py
@@ -106,6 +105,7 @@ frame_path.py
 template.py
 template_data.py
 markdown.py
+markdown_table.py
 markdown_data.py
 markdown_submit_text.py
 markup.py
